Refactor services into core module with improved architecture
